+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.0.2 (YYYY-MM-DD)
+## 1.0.2 (2022-08-05)
 
 ### Breaking Changes
 * None.
@@ -8,7 +8,7 @@
 
 ### Fixed
 * Missing proguard configuration for `CoreErrorUtils`. (Issue [#942](https://github.com/realm/realm-kotlin/issues/942))
-* [Sync] Embedded Objects could not be added to the schema `SyncConfiguration`s. (Issue [#945](https://github.com/realm/realm-kotlin/issues/945)). 
+* [Sync] Embedded Objects could not be added to the schema for `SyncConfiguration`s. (Issue [#945](https://github.com/realm/realm-kotlin/issues/945)). 
 
 ### Compatibility
 * This release is compatible with: